Add files via upload
Implemented a save and load feature. Reworked poison and debugged the hell out of it and is working properly but could use a bit more work on separating the poison damage from the enemy damage on the damage animation. Fixed the double clicking on the buttons. Should not call alerts now on the special button. Re-worked how the game informs the player of the cooldown for the special attack and added animation to the button to let players know it's available to use, Re-wrote drawGrame function for simplicity. Probably more