Add files via upload
Implementation of recruited characters. Have the tents redesigned so they show recruited characters class inside the tent. Re-worked the camp buttons so they are disabled if the player does not have enough gold to purchase camp upgrades but re-active once enough gold is in the camp bank for an upgrade.